pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent_builder"
version = "0.1.0"
description = "MAAP Agent Builder"
authors = [
    {name = "MongoDB", email = "info@mongodb.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    # Core dependencies
    "flask>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "uuid>=1.30",
    "tiktoken>=0.5.1",
    "pipe21>=0.2.0",
    
    # LangChain & related packages
    "langchain>=0.0.340",
    "langchain-core>=0.1.5",
    "langchain-community>=0.0.15",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.0.5",
    "langchain-aws>=0.0.2",
    "langchain-fireworks>=0.1.1",
    "langchain-together>=0.0.2",
    "langchain-cohere>=0.0.1",
    "langchain-huggingface>=0.0.7",
    "langchain-mongodb>=0.0.2",
    "langchain-voyageai>=0.0.1",
    
    # LangGraph
    "langgraph>=0.0.20",
    "langgraph-checkpoint-mongodb",
    
    # MongoDB dependencies
    "pymongo>=4.5.0",
    "certifi>=2023.7.22",
    
    # Vector databases & embedding models
    "sentence-transformers>=2.2.2",
    
    # API & model provider dependencies
    "openai>=1.3.0",
    "anthropic>=0.5.0",
    "cohere>=4.32",
    "boto3",
    "fireworks-ai>=0.7.0",
    "voyageai>=0.1.0",
    
    # Web server
    "gunicorn>=21.2.0",
    
    # Optional: MCP integration
    "langchain-mcp-adapters>=0.1.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    
    # Linting and formatting
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.1",
    "ruff>=0.1.3",
    
    # Documentation
    "sphinx>=7.2.6",
    "sphinx-rtd-theme>=1.3.0",
    
    # Debugging and profiling
    "ipython>=8.16.1",
    "line-profiler>=4.1.1",
    
    # Security scanning
    "bandit>=1.7.5",
]

[project.urls]
"Homepage" = "https://github.com/mongodb/maap-agent-builder"
"Bug Tracker" = "https://github.com/mongodb/maap-agent-builder/issues"

[project.scripts]
agent-builder = "agent_builder.cli:main"

[tool.setuptools]
packages = ["agent_builder"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "B", "I"]
ignore = []

[tool.setuptools.package-data]
agent_builder = ["**/*.yaml", "**/*.txt"]
